Anagram
Given two strings. The task is to check whether the 1st string has a substring which is an anagram of other string.
(An anagram of a string is another string that contains the same characters, only the order of characters can be different. For example, “abcd” and “dabc” are an anagram of each other)