Priority Queue implementation | C++
Priority Queue is an abstract data type which is a collection of elements where the elements have different priority levels. The elements with higher priority is processed before any element of lower priority. If there are elements with the same priority then the element added first in the queue would get processed first. Here this implementation has operations of addition,  deletion,  display, and sorting according to priority .